Refactor game into separate files
